Phase B8: split chat and diagnose handlers into submodule directories
chat.rs (514 lines) -> chat/{mod,streaming}.rs
diagnose.rs (501 lines) -> diagnose/{mod,schema,tool_ceiling,backends}.rs
