+++ crates/defra-agent-cli/src/commands/chat/streaming.rs
@@ -1,244 +1,33 @@
-use std::io::{self, BufRead, Write};
-use std::path::Path;
+use std::io::{self, Write};
 use std::time::Duration;
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 use defra_agent::graphql::escape_graphql_string;
-use serde_json::{json, Value};
+use serde_json::Value;
 
-use crate::cli::args::{ChatArgs, ChatOutputFormat};
-use crate::{
-    create_agent_request, post_graphql, print_json, request_diagnostic_hint, require_non_empty,
-    resolve_home_dir, wait_for_terminal_response, write_json_output_file, RequestSubmitOptions,
-    SubmittedRequest, DEFAULT_AGENT_NAME, DEFAULT_HTTP_PORT,
-};
+use crate::{post_graphql, request_diagnostic_hint};
 
-pub(crate) async fn chat(args: ChatArgs) -> Result<()> {
-    let home_dir = resolve_home_dir(args.home.as_deref());
-    let runtime_state = crate::read_runtime_state(&home_dir)?;
-    let init_config = crate::read_init_config(&home_dir)?;
-    let graphql = args
-        .graphql
-        .clone()
-        .or_else(|| runtime_state.as_ref().map(|state| state.graphql.clone()))
-        .unwrap_or_else(|| format!("http://127.0.0.1:{DEFAULT_HTTP_PORT}/api/v0/graphql"));
-    let agent_name = args
-        .agent_name
-        .clone()
-        .or_else(|| runtime_state.as_ref().map(|state| state.agent_name.clone()))
-        .or_else(|| init_config.as_ref().map(|config| config.agent_name.clone()))
-        .unwrap_or_else(|| DEFAULT_AGENT_NAME.to_string());
-    let agent_did = args
-        .agent_did
-        .clone()
-        .or_else(|| runtime_state.as_ref().map(|state| state.agent_did.clone()))
-        .or_else(|| init_config.as_ref().map(|config| config.agent_did.clone()))
-        .unwrap_or_else(|| format!("did:defra-agent:{agent_name}"));
-    let session_id = args
-        .session_id
-        .clone()
-        .unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
+use super::SubmittedRequest;
 
-    if let Some(message) = resolve_chat_message(&args.message, args.message_file.as_deref())? {
-        match args.output_format {
-            ChatOutputFormat::Text => {
-                let (_submitted, response) = submit_chat_turn(
-                    &graphql,
-                    &agent_did,
-                    &session_id,
-                    args.behavior_id.as_deref(),
-                    &message,
-                    args.timeout_secs,
-                    args.poll_secs,
-                )
-                .await?;
-                if let Some(path) = args.output_file.as_deref() {
-                    write_text_output_file(path, response_text_content(&response))?;
-                }
-            }
-            ChatOutputFormat::Json => {
-                let output = submit_chat_turn_json(
-                    &graphql,
-                    &agent_did,
-                    &session_id,
-                    args.behavior_id.as_deref(),
-                    &message,
-                    args.timeout_secs,
-                    args.poll_secs,
-                )
-                .await?;
-                print_json(&output)?;
-                if let Some(path) = args.output_file.as_deref() {
-                    write_json_output_file(path, &output)?;
-                }
-            }
-        }
-        return Ok(());
-    }
-
-    if args.output_format != ChatOutputFormat::Text {
-        anyhow::bail!("interactive chat only supports --output-format text");
-    }
-    if let Some(path) = args.output_file.as_deref() {
-        anyhow::bail!(
-            "--output-file {} requires a one-shot message via MESSAGE or --message-file",
-            path.display()
-        );
-    }
-
-    let stdin = io::stdin();
-    let mut lines = stdin.lock().lines();
-    let mut stdout = io::stdout();
-    loop {
-        write!(stdout, "> ")?;
-        stdout.flush()?;
-        let Some(line) = lines.next() else {
-            break;
-        };
-        let line = line?;
-        let trimmed = line.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if matches!(trimmed, "/exit" | "/quit" | "exit" | "quit") {
-            break;
-        }
-
-        submit_chat_turn(
-            &graphql,
-            &agent_did,
-            &session_id,
-            args.behavior_id.as_deref(),
-            trimmed,
-            args.timeout_secs,
-            args.poll_secs,
-        )
-        .await?;
-    }
-
-    Ok(())
-}
-
-async fn submit_chat_turn(
-    graphql: &str,
-    agent_did: &str,
-    session_id: &str,
-    behavior_id: Option<&str>,
-    content: &str,
-    timeout_secs: u64,
-    poll_secs: u64,
-) -> Result<(SubmittedRequest, Value)> {
-    let existing_tool_calls = load_existing_tool_call_keys(graphql, session_id).await?;
-    let submitted = create_agent_request(
-        graphql,
-        agent_did,
-        content,
-        Some(session_id),
-        behavior_id,
-        RequestSubmitOptions::default(),
-    )
-    .await?;
-    let response = stream_turn_progress(
-        graphql,
-        &submitted,
-        existing_tool_calls,
-        timeout_secs,
-        poll_secs,
-    )
-    .await?;
-    Ok((submitted, response))
-}
-
-async fn submit_chat_turn_json(
-    graphql: &str,
-    agent_did: &str,
-    session_id: &str,
-    behavior_id: Option<&str>,
-    content: &str,
-    timeout_secs: u64,
-    poll_secs: u64,
-) -> Result<Value> {
-    let submitted = create_agent_request(
-        graphql,
-        agent_did,
-        content,
-        Some(session_id),
-        behavior_id,
-        RequestSubmitOptions::default(),
-    )
-    .await?;
-    let response =
-        wait_for_terminal_response(graphql, &submitted.request_id, timeout_secs, poll_secs)
-            .await
-            .with_context(|| format!("waiting for AgentResponse {}", submitted.request_id))?;
-    Ok(chat_turn_output(&submitted, response))
-}
-
-fn resolve_chat_message(message: &[String], message_file: Option<&Path>) -> Result<Option<String>> {
-    if !message.is_empty() && message_file.is_some() {
-        anyhow::bail!("provide either MESSAGE or --message-file, not both");
-    }
-    if !message.is_empty() {
-        return Ok(Some(
-            require_non_empty("message", &message.join(" "))?.to_string(),
-        ));
-    }
-    if let Some(path) = message_file {
-        let message = std::fs::read_to_string(path)
-            .with_context(|| format!("reading chat message from {}", path.display()))?;
-        return Ok(Some(
-            require_non_empty("message-file", &message)?.to_string(),
-        ));
-    }
-    Ok(None)
-}
-
-fn write_text_output_file(path: &Path, content: &str) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)
-            .with_context(|| format!("creating output directory {}", parent.display()))?;
-    }
-    std::fs::write(path, content)
-        .with_context(|| format!("writing text output file {}", path.display()))?;
-    Ok(())
-}
-
-fn response_text_content(response: &Value) -> &str {
-    response
-        .get("content")
-        .and_then(Value::as_str)
-        .unwrap_or("")
-}
-
-fn chat_turn_output(submitted: &SubmittedRequest, response: Value) -> Value {
-    json!({
-        "request_id": submitted.request_id,
-        "session_id": submitted.session_id,
-        "agent_did": submitted.agent_did,
-        "behavior_id": submitted.behavior_id,
-        "response": response,
-    })
+#[derive(Debug, Clone)]
+pub(super) struct ChatTurnProgress {
+    pub(super) content: String,
+    pub(super) reasoning: String,
+    pub(super) error_message: Option<String>,
+    pub(super) progress_seq: u64,
+    pub(super) status: String,
 }
 
 #[derive(Debug, Clone)]
-struct ChatTurnProgress {
-    content: String,
-    reasoning: String,
-    error_message: Option<String>,
-    progress_seq: u64,
-    status: String,
+pub(super) struct ToolCallProgress {
+    pub(super) tool_call_key: String,
+    pub(super) tool_name: String,
+    pub(super) status: String,
+    pub(super) args: String,
+    pub(super) result: String,
 }
 
-#[derive(Debug, Clone)]
-struct ToolCallProgress {
-    tool_call_key: String,
-    tool_name: String,
-    status: String,
-    args: String,
-    result: String,
-}
-
-fn chat_progress_query(request_id: &str, session_id: &str) -> String {
+pub(super) fn chat_progress_query(request_id: &str, session_id: &str) -> String {
     format!(
         r#"{{
             AgentResponse(
@@ -273,7 +62,7 @@ fn chat_progress_query(request_id: &str, session_id: &str) -> String {
     )
 }
 
-async fn load_existing_tool_call_keys(
+pub(super) async fn load_existing_tool_call_keys(
     graphql: &str,
     session_id: &str,
 ) -> Result<std::collections::BTreeMap<String, String>> {
@@ -309,7 +98,7 @@ async fn load_existing_tool_call_keys(
         .collect())
 }
 
-async fn stream_turn_progress(
+pub(super) async fn stream_turn_progress(
     graphql: &str,
     submitted: &SubmittedRequest,
     mut known_tool_calls: std::collections::BTreeMap<String, String>,
@@ -429,7 +218,7 @@ async fn stream_turn_progress(
     }
 }
 
-fn decode_chat_turn_progress(row: &Value) -> Option<ChatTurnProgress> {
+pub(super) fn decode_chat_turn_progress(row: &Value) -> Option<ChatTurnProgress> {
     Some(ChatTurnProgress {
         content: row.get("content")?.as_str()?.to_string(),
         reasoning: row
@@ -446,7 +235,7 @@ fn decode_chat_turn_progress(row: &Value) -> Option<ChatTurnProgress> {
     })
 }
 
-fn decode_tool_call_progress(row: &Value) -> Option<ToolCallProgress> {
+pub(super) fn decode_tool_call_progress(row: &Value) -> Option<ToolCallProgress> {
     Some(ToolCallProgress {
         tool_call_key: row.get("tool_call_key")?.as_str()?.to_string(),
         tool_name: row.get("tool_name")?.as_str()?.to_string(),
@@ -464,7 +253,7 @@ fn decode_tool_call_progress(row: &Value) -> Option<ToolCallProgress> {
     })
 }
 
-fn format_tool_progress_line(tool: &ToolCallProgress) -> String {
+pub(super) fn format_tool_progress_line(tool: &ToolCallProgress) -> String {
     match tool.status.as_str() {
         "completed" => match preview_compact_text(&tool.result) {
             Some(result) => format!(
@@ -493,13 +282,13 @@ fn format_tool_progress_line(tool: &ToolCallProgress) -> String {
     }
 }
 
-fn format_tool_args_preview(value: &str) -> String {
+pub(super) fn format_tool_args_preview(value: &str) -> String {
     preview_compact_text(value)
         .map(|preview| format!("({preview})"))
         .unwrap_or_default()
 }
 
-fn preview_compact_text(value: &str) -> Option<String> {
+pub(super) fn preview_compact_text(value: &str) -> Option<String> {
     let compact = value.split_whitespace().collect::<Vec<_>>().join(" ");
     let trimmed = compact.trim();
     if trimmed.is_empty() {
