+++ crates/defra-agent-cli/src/commands/diagnose/mod.rs
@@ -1,0 +1,196 @@
+mod backends;
+mod schema;
+mod tool_ceiling;
+
+use anyhow::Result;
+use serde_json::{json, Value};
+
+use crate::cli::args::{DiagnoseArgs, P2pTransportArg};
+use crate::shared::{ConfigExportBundle, StoredInitConfig};
+use crate::{
+    build_config_export_bundle, graphql_endpoint_available, print_json, read_init_config,
+    read_runtime_state, resolve_agent_did, resolve_config_access, resolve_home_dir,
+    CONFIG_EXPORT_FORMAT,
+};
+
+use backends::diagnose_backends;
+use schema::{diagnose_schema_presence, load_runtime_row};
+use tool_ceiling::diagnose_tool_ceiling;
+
+pub(crate) async fn diagnose(args: DiagnoseArgs) -> Result<()> {
+    let home_dir = resolve_home_dir(args.home.as_deref());
+    let init_config = read_init_config(&home_dir)?;
+    let runtime_state = read_runtime_state(&home_dir)?;
+    let graphql = args
+        .graphql
+        .clone()
+        .or_else(|| runtime_state.as_ref().map(|state| state.graphql.clone()));
+    let graphql_reachable = match graphql.as_deref() {
+        Some(endpoint) => graphql_endpoint_available(endpoint).await,
+        None => false,
+    };
+    let agent_did = resolve_agent_did(args.home.as_deref(), args.agent_did.as_deref())?;
+    let (access, _) =
+        resolve_config_access(args.home.as_deref(), args.graphql.as_deref(), false).await?;
+
+    let schema_checks = diagnose_schema_presence(&access).await;
+    let bundle_result = build_config_export_bundle(&access, &agent_did).await;
+    let config_load_error = bundle_result.as_ref().err().map(ToString::to_string);
+    let bundle = bundle_result.unwrap_or_else(|_| ConfigExportBundle {
+        format: CONFIG_EXPORT_FORMAT.to_string(),
+        agent_did: agent_did.clone(),
+        exported_at: chrono::Utc::now().to_rfc3339(),
+        access_mode: access.mode().to_string(),
+        agent_principal: None,
+        agent_behaviors: Vec::new(),
+        tool_selections: Vec::new(),
+        inference_backends: Vec::new(),
+        inference_profiles: Vec::new(),
+        tool_service_registries: Vec::new(),
+        scheduled_tasks: Vec::new(),
+    });
+    let runtime_row = match load_runtime_row(&access, &agent_did).await {
+        Ok(Some(row)) => row,
+        Ok(None) => Value::Null,
+        Err(error) => json!({
+            "error": error.to_string(),
+        }),
+    };
+
+    let behavior_ids = bundle
+        .agent_behaviors
+        .iter()
+        .filter_map(|row| {
+            row.get("behavior_id")
+                .and_then(Value::as_str)
+                .map(ToOwned::to_owned)
+        })
+        .collect::<std::collections::BTreeSet<_>>();
+    let default_behavior_id = bundle
+        .agent_principal
+        .as_ref()
+        .and_then(|row| row.get("default_behavior_id"))
+        .and_then(Value::as_str)
+        .map(ToOwned::to_owned);
+    let default_behavior_check = match default_behavior_id.as_deref() {
+        Some(behavior_id) if behavior_ids.contains(behavior_id) => json!({
+            "ok": true,
+            "default_behavior_id": behavior_id,
+        }),
+        Some(behavior_id) => json!({
+            "ok": false,
+            "default_behavior_id": behavior_id,
+            "error": format!("default behavior {} is not present in AgentBehavior documents", behavior_id),
+        }),
+        None => json!({
+            "ok": false,
+            "error": format!("AgentPrincipal {} is missing or has no default_behavior_id", agent_did),
+        }),
+    };
+    let tool_ceiling_check = diagnose_tool_ceiling(init_config.as_ref());
+    let backend_reports = diagnose_backends(&bundle).await;
+    let matching_runtime_state = runtime_state.as_ref().filter(|state| {
+        graphql
+            .as_deref()
+            .is_some_and(|endpoint| endpoint == state.graphql)
+    });
+    let p2p_status = match graphql.as_deref().filter(|_| graphql_reachable) {
+        Some(endpoint) => {
+            crate::commands::p2p::load_live_http_p2p_status(args.home.as_deref(), endpoint).await
+        }
+        None => crate::commands::p2p::persisted_p2p_status(matching_runtime_state),
+    };
+    let p2p_transport = p2p_status
+        .get("p2p_transport")
+        .and_then(Value::as_str)
+        .unwrap_or(P2pTransportArg::None.as_str());
+    let p2p_peer_id = p2p_status
+        .get("p2p_peer_id")
+        .and_then(Value::as_str)
+        .map(ToOwned::to_owned);
+    let p2p_connected_peers = p2p_status
+        .get("p2p_connected_peers")
+        .and_then(Value::as_array)
+        .map(|rows| rows.len())
+        .unwrap_or(0);
+    let p2p_error = p2p_status
+        .get("p2p_error")
+        .and_then(Value::as_str)
+        .map(ToOwned::to_owned);
+    let p2p_ok = if p2p_transport == P2pTransportArg::None.as_str() {
+        true
+    } else {
+        p2p_peer_id.is_some() && p2p_error.is_none()
+    };
+    let schemas_ok = schema_checks
+        .iter()
+        .filter(|check| check.get("required_for_config").and_then(Value::as_bool) == Some(true))
+        .all(|check| check.get("ok").and_then(Value::as_bool) == Some(true));
+    let backends_ok = backend_reports
+        .iter()
+        .all(|check| check.get("ok").and_then(Value::as_bool) == Some(true));
+    let default_behavior_ok = default_behavior_check
+        .get("ok")
+        .and_then(Value::as_bool)
+        .unwrap_or(false);
+    let tool_ceiling_ok = tool_ceiling_check
+        .get("ok")
+        .and_then(Value::as_bool)
+        .unwrap_or(false);
+    let principal_present = bundle.agent_principal.is_some();
+    let status = if schemas_ok
+        && principal_present
+        && default_behavior_ok
+        && tool_ceiling_ok
+        && backends_ok
+        && p2p_ok
+        && config_load_error.is_none()
+    {
+        "ok"
+    } else {
+        "degraded"
+    };
+
+    let mut output = json!({
+        "status": status,
+        "home": home_dir,
+        "agent_did": agent_did,
+        "access_mode": access.mode(),
+        "graphql": graphql,
+        "graphql_reachable": graphql_reachable,
+        "runtime": runtime_row,
+        "p2p": p2p_status,
+        "checks": {
+            "schemas": schema_checks,
+            "config_documents_loadable": {
+                "ok": config_load_error.is_none(),
+                "error": config_load_error,
+            },
+            "agent_principal_present": principal_present,
+            "default_behavior": default_behavior_check,
+            "tool_ceiling": tool_ceiling_check,
+            "backends": backend_reports,
+            "p2p": {
+                "ok": p2p_ok,
+                "transport": p2p_transport,
+                "peer_id": p2p_peer_id,
+                "connected_peer_count": p2p_connected_peers,
+                "error": p2p_error,
+            },
+        },
+        "config_counts": {
+            "agent_behaviors": bundle.agent_behaviors.len(),
+            "tool_selections": bundle.tool_selections.len(),
+            "inference_backends": bundle.inference_backends.len(),
+            "inference_profiles": bundle.inference_profiles.len(),
+            "tool_service_registries": bundle.tool_service_registries.len(),
+            "scheduled_tasks": bundle.scheduled_tasks.len(),
+        },
+    });
+    if let Some(map) = output.as_object_mut() {
+        let p2p_value = map.get("p2p").cloned().unwrap_or(Value::Null);
+        crate::commands::p2p::flatten_p2p_fields(map, &p2p_value);
+    }
+    print_json(&output)?;
+    Ok(())
+}
